+++ projects/02_architecture_comparison_lab/mvc/Sources/PFMVCStudyTaskStore.h
@@ -5,6 +5,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface PFMVCStudyTaskStore : NSObject
 - (NSArray<PFMVCStudyTask *> *)defaultTasks;
+- (void)simulateNextSaveFailure;
 - (nullable NSString *)saveTasks:(NSArray<PFMVCStudyTask *> *)tasks error:(NSError **)error;
 - (nullable NSArray<PFMVCStudyTask *> *)loadTasksFromPath:(NSString *)path error:(NSError **)error;
 @end