+++ projects/02_architecture_comparison_lab/mvp/Sources/PFMVPStudyTaskStore.h
@@ -5,6 +5,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface PFMVPStudyTaskStore : NSObject
 - (NSArray<PFMVPStudyTask *> *)defaultTasks;
+- (void)simulateNextSaveFailure;
 - (nullable NSString *)saveTasks:(NSArray<PFMVPStudyTask *> *)tasks error:(NSError **)error;
 - (nullable NSArray<PFMVPStudyTask *> *)loadTasksFromPath:(NSString *)path error:(NSError **)error;
 @end