+++ projects/02_architecture_comparison_lab/mvvm/Sources/PFMVVMStudyTaskStore.h
@@ -5,6 +5,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface PFMVVMStudyTaskStore : NSObject
 - (NSArray<PFMVVMStudyTask *> *)defaultTasks;
+- (void)simulateNextSaveFailure;
 - (nullable NSString *)saveTasks:(NSArray<PFMVVMStudyTask *> *)tasks error:(NSError **)error;
 - (nullable NSArray<PFMVVMStudyTask *> *)loadTasksFromPath:(NSString *)path error:(NSError **)error;
 @end