+++ projects/02_architecture_comparison_lab/viper/Sources/PFVIPERStudyTaskStore.h
@@ -5,6 +5,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface PFVIPERStudyTaskStore : NSObject
 - (NSArray<PFVIPERStudyTask *> *)defaultTasks;
+- (void)simulateNextSaveFailure;
 - (nullable NSString *)saveTasks:(NSArray<PFVIPERStudyTask *> *)tasks error:(NSError **)error;
 - (nullable NSArray<PFVIPERStudyTask *> *)loadTasksFromPath:(NSString *)path error:(NSError **)error;
 @end